+++ spec/image_optim/worker/optipng_spec.rb
@@ -59,8 +59,8 @@ describe ImageOptim::Worker::Optipng do
   describe '#optimized?' do
     let(:src){ instance_double(ImageOptim::Path, src_options) }
     let(:dst){ instance_double(ImageOptim::Path, dst_options) }
-    let(:src_options){ {:size? => 10, :size => 10} }
-    let(:dst_options){ {:size? => 9, :size => 9} }
+    let(:src_options){ {:size => 10} }
+    let(:dst_options){ {:size? => 9} }
     let(:instance){ described_class.new(ImageOptim.new, instance_options) }
     let(:instance_options){ {} }
 
@@ -75,7 +75,7 @@ describe ImageOptim::Worker::Optipng do
       end
 
       context 'when dst is not empty' do
-        let(:dst_options){ {:size? => 20, :size => 20} }
+        let(:dst_options){ {:size? => 20} }
         it{ is_expected.to be_truthy }
       end
     end
@@ -89,12 +89,12 @@ describe ImageOptim::Worker::Optipng do
       end
 
       context 'when dst is greater than or equal to src' do
-        let(:dst_options){ {:size? => 10, :size => 10} }
+        let(:dst_options){ {:size? => 10} }
         it{ is_expected.to be_falsy }
       end
 
       context 'when dst is less than src' do
-        let(:dst_options){ {:size? => 9, :size => 9} }
+        let(:dst_options){ {:size? => 9} }
         it{ is_expected.to be_truthy }
       end
     end