+++ lib/image_optim.rb
@@ -5,7 +5,6 @@ require 'image_optim/handler'
 require 'image_optim/image_meta'
 require 'image_optim/optimized_path'
 require 'image_optim/path'
-require 'image_optim/railtie' if defined?(Rails)
 require 'image_optim/worker'
 require 'in_threads'
 require 'shellwords'