+++ CHANGELOG.markdown
@@ -3,6 +3,7 @@
 ## unreleased
 
 * Add proper handling of `ImageOptim.respond_to?` [@toy](https://github.com/toy)
+* Fix an issue not working OptiPNG `interlace` option [#136](https://github.com/toy/image_optim/pull/136) [@mrk21](https://github.com/mrk21)
 
 ## v0.23.0 (2016-07-17)