+++ CHANGELOG.markdown
@@ -2,6 +2,8 @@
 
 ## unreleased
 
+## v0.24.0 (2016-08-14)
+
 * Rails image assets optimization is extracted into [image\_optim\_rails gem](https://github.com/toy/image_optim_rails) [#127](https://github.com/toy/image_optim/issues/127) [@toy](https://github.com/toy)
 * Add proper handling of `ImageOptim.respond_to?` [@toy](https://github.com/toy)
 * Fix an issue not working OptiPNG `interlace` option [#136](https://github.com/toy/image_optim/pull/136) [@mrk21](https://github.com/mrk21)